v1.0.31121.1442
When ShouldExpand=true and tabs are added and removed dynamically, the
tabs will now correctly resize.

This has been achieved by forcing a second measure pass when the tab
widths are checked.

There's a question mark over the behaviour when the number of tabs
decreases and the currently selected tab is greater than the eventual
number of tabs - however, it doesn't crash.

The example app has been updated to demonstrate the number of tabs
changing dynamically.