1.1.31128.1503 (released to nuget)
Massive change - added new interface ITabProvider through which all tab
creation now goes; original text and icon tab functionality now
implemented through first-class implementations of the ITabProvider.
Developers implement ITabProvider on their pager adapter to enable the
functionality.  An abstract base class is provided to simplify this:
FragmentTabProviderPagerAdapterBase, which you should be able to simply
swap in for FragmentPagerAdapter.

This adds view recycling for tabs and partial update notifications (for
when a tab's content changes and the tabstrip needs to re-layout).

The example has been updated to demonstrate an implementation of
ITabProvider that embeds a progress bar in the tab with a textview -
this is done as a first-class implementation of ITabProvider rather than
utilising the FragmentTabProviderPagerAdater class.

Note that tab state management (i.e. data that influences the content of
the tabview) is in particular very important, and a couple of ways of
managing this are demonstrated.